pyproject.toml
[project]
name = "textgen-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=2.1",
    "groq>=0.1",
    "python-dotenv>=1.0.0",
]
